rename header + icons
