+++ .env
@@ -1,2 +1,3 @@
-REACT_APP_BASE_URL_DEVELOPMENT=http://localhost:3001
+#REACT_APP_BASE_URL_DEVELOPMENT=http://localhost:3001
+REACT_APP_BASE_URL_DEVELOPMENT=https://us-central1-study-app1-8cc56.cloudfunctions.net/app
 REACT_APP_BASE_URL_PRODUCTION=https://us-central1-study-app1-8cc56.cloudfunctions.net/app
