+++ src/App.js
@@ -3,7 +3,7 @@ import { HashRouter as Router, Route, Routes } from 'react-router-dom';
 import { CreatePost } from './posts';
 import { Home } from './home';
 import { Login } from './auth';
-import { NavBar } from './header/MainNavbar/NavBar';
+import { Header } from './header/';
 import { ErrorBoundary } from './errorBoundary';
 // import { AnimalsDropdown } from './sandbox/dropdown/AnimalsDrop';
 
@@ -16,7 +16,7 @@ export function App() {
 
   return (
     <Router basename="/">
-      <NavBar />
+      <Header />
       <ErrorBoundary>
         <Routes>
           <Route exact path="/" element={<Home />} />
