+++ src/header/Header.module.css
@@ -86,6 +86,7 @@
   justify-content: center;
   column-gap: 10px;
   margin-left: auto;
+  margin-right: auto;
 }
 
 .newsLink {
@@ -127,10 +128,10 @@
 }
 
 .authImg {
-  width: 38px;
-  height: 38px;
+  width: 48px;
+  height: 48px;
   border: 2px solid #dddddd;
-  border-radius: 64px;
+  border-radius: 50%;
 }
 
 .blog {
