+++ src/header/Logo/Logo.module.css
@@ -1,0 +1,7 @@
+.logoWrap {
+  height: 100%;
+}
+
+.logoImg {
+  height: 100%;
+}
