+++ src/header/Logo/index.js
@@ -1,0 +1,1 @@
+export { Logo } from './Logo'
