+++ src/header/Profile/Profile.js
@@ -1,10 +1,10 @@
 import { observer } from 'mobx-react';
 import React from 'react';
-import styles from '../Navbar.module.css';
+import styles from '../Header.module.css';
 import { authService } from '../../auth/auth.service';
 
-export const RightNavbar = observer(
-  class RightNavbar extends React.Component {
+export const Profile = observer(
+  class Profile extends React.Component {
     render() {
       const ImageSrc = authService.photoSrc;
       return (
