+++ src/header/index.js
@@ -1,0 +1,1 @@
+export { Header } from './Header';
